Add new option to limit duplicate query tracebacks
By default, this plugin shows only the first traceback for a duplicated
query. The option introduced in this commit allows the user to specify a
maximum number of duplicate queries for which tracebacks should be
printed. This should make it easier to track down where duplicate query
calls may be occurring in the application.